Adaugare hook custom pentru clienti
